Ask React Native to resolve props
Apparently, some props need to be resolved by React Native. I noticed this with RCTImageView's src prop. This code is kind of a mess because I don't know the contract about which props need to be resolved via React Native, which don't, and whether or not I need to combine information from before and after the resolution. Once I understand the contract, this code can be cleaned up.